Add bin downloader to utils
