+++ lib/allure_ruby_commons/result_utils.rb
@@ -27,6 +27,10 @@ module Allure
     LANGUAGE_LABEL_NAME = "language"
 
     class << self
+      def timestamp(time = nil)
+        ((time || Time.now).to_f * 1000).to_i
+      end
+
       def thread_label
         Label.new(THREAD_LABEL_NAME, Thread.current.object_id)
       end
